refactor: PC View, Mobile View 파일 통합
