--- src/pages/Club/NewClubRecruitment/components/MobileView/index.tsx
@@ -1,8 +1,0 @@
-export default function MobileView() {
-  return (
-    <div>
-      <h1>Mobile View</h1>
-      <p>This is the mobile view for the new club recruitment page.</p>
-    </div>
-  );
-}
