--- src/pages/Club/NewClubRecruitment/components/PCView/NewClubRecruitmentPCView.module.scss
@@ -1,224 +1,0 @@
-@use "src/utils/scss/media" as media;
-
-.layout {
-  width: 100%;
-  height: 100%;
-  box-sizing: border-box;
-  display: flex;
-  justify-content: center;
-  font-family: Pretendard, sans-serif;
-}
-
-.container {
-  width: 1136px;
-  padding-top: 56px;
-  padding-bottom: 60px;
-}
-
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  border-bottom: #175c8e 2px solid;
-  padding-bottom: 24px;
-
-  &__title {
-    color: #175c8e;
-    font-size: 20px;
-    font-weight: 500;
-  }
-
-  &__button-container {
-    display: flex;
-    align-items: center;
-    gap: 24px;
-  }
-
-  &__button {
-    display: flex;
-    align-items: center;
-    padding: 9px 20px;
-    background-color: #175c8e;
-    font-size: 14px;
-    color: #fff;
-    border-radius: 4px;
-    line-height: 160%;
-  }
-}
-
-.content {
-  padding-top: 20px;
-  display: flex;
-  justify-content: space-between;
-}
-
-.form-left {
-  display: flex;
-  flex-direction: column;
-  gap: 23px;
-}
-
-.form {
-  &__item {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 12px;
-    width: 550px;
-  }
-
-  &__item-title {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    > p {
-      font-size: 20px;
-      font-weight: 700;
-    }
-  }
-
-  &__label {
-    font-weight: 500;
-    font-size: 22px;
-  }
-
-  &__input {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-  }
-
-  &__date-input {
-    width: 100%;
-    padding: 12px 16px;
-    border: 1px solid #eee;
-    border-radius: 8px;
-    font-size: 16px;
-    line-height: 160%;
-  }
-
-  &__checkbox {
-    width: 18px;
-    height: 18px;
-    margin: 0;
-  }
-
-  &__textarea {
-    width: 100%;
-    height: 68px;
-    box-sizing: border-box;
-    padding: 8px 16px;
-    border-radius: 4px;
-    resize: none;
-    border: 1px solid #6dbbdd;
-    background-color: #f5f5f5;
-    font-family: Pretendard, sans-serif;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 1.6;
-
-    ::placeholder {
-      color: #4b4b4b;
-    }
-  }
-}
-
-.form-image {
-  &__input {
-    display: none;
-  }
-
-  &__label {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 251px 142px;
-    background-color: #eee;
-    border-radius: 16px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-
-    &--drag-over {
-      background-color: #d0d0d0;
-    }
-
-    &:hover {
-      background-color: #e0e0e0;
-    }
-
-    &--error {
-      border: 1px solid #f7941e;
-    }
-  }
-
-  &__label-text {
-    color: #727272;
-    text-align: center;
-    font-family: Pretendard, sans-serif;
-    font-size: 18px;
-    font-weight: 500;
-    line-height: 160%;
-  }
-
-  &__preview {
-    width: 500px;
-    overflow: hidden;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    gap: 5px;
-  }
-
-  &__img {
-    min-width: 400px;
-    max-width: 100%;
-    max-height: 100%;
-    object-fit: contain;
-
-    &__box {
-      width: 100%;
-      display: flex;
-      justify-content: flex-end;
-    }
-  }
-
-  &__edit-button {
-    padding: 5px 12px;
-    border-radius: 4px;
-    background-color: #175c8e;
-    color: #fff;
-    font-family: Pretendard, sans-serif;
-    font-size: 20px;
-    font-weight: 500;
-    white-space: nowrap;
-
-    &__box {
-      width: 100%;
-      display: flex;
-      justify-content: flex-end;
-    }
-  }
-
-  &__description {
-    color: #727272;
-    font-size: 12px;
-    line-height: 160%;
-  }
-}
-
-.date-picker-button {
-  width: 250px;
-  padding: 4px;
-  background-color: #fafafa;
-  border: 1px solid #e1e1e1;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0 0 0 / 40%), 0 1px 1px rgba(0 0 0 / 10%);
-  font-family: pretendard, sans-serif;
-  font-size: 18px;
-  font-weight: 500;
-  line-height: 160%;
-}
